frFont: use typed glyph-node return alias

+++ src/Event/mt_evtMain.c
@@ -650,7 +650,7 @@ void FUN_00366210(u32 param_1,int param_2);
 void FUN_00366330(u32 param_1,int param_2);
 void FUN_003663f0(u32 param_1,int param_2);
 void FUN_00366660(u32 param_1,u64 param_2,u64 param_3,u64 param_4);
-void FUN_00366970(u64 param_1,u32 param_2);
+void FUN_00366970(u32 param_1,u32 param_2);
 void FUN_00366ac0(u32 param_1,u64 param_2);
 void FUN_0036d030(int param_1);
 void FUN_0036efb0(void);
@@ -6700,7 +6700,7 @@ LAB_00366928:;
 // FUN_00366970 NONMATCHING
 
 
-void FUN_00366970(u64 param_1,u32 param_2)
+void FUN_00366970(u32 param_1,u32 param_2)
